[JENKINS-37129] unclassified method vs MethodMissingException
As an user, I don't always precisely remember what methods an object
has, an I often end up making mistakes. From workflow-cps perspective,
this results in trying to call a method that does not exist.

This must result in MethodMissingException, just like normal Groovy
interpretor. The code currently makes this a RejectedAccessException,
making the user believe that the method exists but the access it not
allowed.