Increasing test coverage of Groovy syntax
With some of the regressions recently, we probably would have caught
them if we'd had better coverage of the language as a whole. So, let's
add more coverage. This has already uncovered one bug, JENKINS-46327,
and at least one fairly simple operator call that needed a
whitelisting (`3.0 * 5.0`). More to come...