[FIXED JENKINS-46764] Trim leading/trailing ws from signatures
If a signature string starts with whitespace currently, that results
in rejecting all signatures that are approved in that StaticWhitelist,
due to the IOException thrown when parsing. So let's trim away said
whitespace, and also ensure that adding a signature trims it as well.

Perhaps a more comprehensive solution is needed for not barfing out
everything on one failed-to-parse signature, though.