Add various Run methods to the whitelist
This adds several of the readonly methods so that another job can access
them. For example a job that takes a parameter or input step of
selecting a job would then be able to access the result, changesets, URL
and so on of the other job.

The methods setDisplayName and setDescription are included so that
another job can use them to annotate/update as a simple promoted builds
style implementation.