Add proper support for propertyMissing setter and getter
85d16bd broke the setter version of propertyMissing method. Previously,
Groovys synthetic setProperty() method looked for propertyMissing()
methods and called it if it was defined. After 85d16bd this is no
longer supported since calls to the synthetic version of setProperty()
is prohibited.

This commit addresses this issue by explicitly checking for
propertyMissing() methods, both in the set and get cases.