Fix reloading configuration from disk
This makes the `load()` method work properly when called after startup, in
case of a reload configuration from disk for example.

Functionally, the `reconfigure()` method needs to be called implicitly when calling
`load()`, because it updates a snapshot state that is used by running builds.
If the persisted state has changed on disk then loaded, the snapshot needs to be updated accordingly.

Without this patch, the `ScriptApproval` state is correct, but anything
that has changed since last time it was loaded is not reflected on
running builds.